+++ internal/modules/services/topology/dto.go
@@ -5,11 +5,11 @@ package topology
 // latency_state.
 type nodeAggRow struct {
 	ServiceName  string  `ch:"service"`
-	RequestCount int64   `ch:"request_count"`
-	ErrorCount   int64   `ch:"error_count"`
-	P50Ms        float64 `ch:"p50_ms"`
-	P95Ms        float64 `ch:"p95_ms"`
-	P99Ms        float64 `ch:"p99_ms"`
+	RequestCount uint64  `ch:"request_count"`
+	ErrorCount   uint64  `ch:"error_count"`
+	P50Ms        float32 `ch:"p50_ms"`
+	P95Ms        float32 `ch:"p95_ms"`
+	P99Ms        float32 `ch:"p99_ms"`
 }
 
 // edgeAggRow is scanned from the per-(service → peer_service) edge
@@ -17,8 +17,8 @@ type nodeAggRow struct {
 type edgeAggRow struct {
 	Source     string  `ch:"source"`
 	Target     string  `ch:"target"`
-	CallCount  int64   `ch:"call_count"`
-	ErrorCount int64   `ch:"error_count"`
-	P50Ms      float64 `ch:"p50_ms"`
-	P95Ms      float64 `ch:"p95_ms"`
+	CallCount  uint64  `ch:"call_count"`
+	ErrorCount uint64  `ch:"error_count"`
+	P50Ms      float32 `ch:"p50_ms"`
+	P95Ms      float32 `ch:"p95_ms"`
 }
